Moved logic for start of pipes using threads helpers.py
